kyhton.kompress: move to core (with a fallback, used in promnesia)
