+++ my/kython/README
@@ -1,1 +1,1 @@
-vendorized kython (https://github.com/karlicoss/kython) stuff
+deprecated, please use my.core directly
