+++ tests/misc.py
@@ -5,7 +5,7 @@ import lzma
 import io
 import zipfile
 
-from my.kython.kompress import kopen, kexists, CPath
+from my.core.kompress import kopen, kexists, CPath
 
 def test_kopen(tmp_path: Path) -> None:
     "Plaintext handled transparently"
